cleanup:

+++ cleanup.py
@@ -17,7 +17,10 @@ def main():
     cleanup_deleted_files()
 
 def cleanup_zero_size_files():
-    syslog.syslog(syslog.LOG_INFO, "Delete any 0 byte info files from the movie and tv folders.")
+    counter = 0
+    delete_threshhold = arrow.now().shift(days=-1)
+
+    syslog.syslog(syslog.LOG_INFO, "Delete any 0 byte files from the movie and tv folders.")
 
     target_size=0
 
@@ -26,8 +29,13 @@ def cleanup_zero_size_files():
             for file in files: 
                 path = os.path.join(dirpath, file)
                 if os.stat(path).st_size <= target_size:
-                    syslog.syslog(syslog.LOG_INFO, "Deleting {0}".format(path))
-                    os.remove(path)
+                    mod_time = arrow.get(os.stat(path).st_ctime)
+                    if mod_time < delete_threshhold:
+                        syslog.syslog(syslog.LOG_INFO, "Deleting {0}".format(path))
+                        os.remove(path)
+                        counter += 1
+
+    syslog.syslog(syslog.LOG_INFO, "Deleted {0} empty files from the movie and tv folders.")
 
 def cleanup_deleted_files():
     delete_threshhold = arrow.now().shift(days=-30)
@@ -56,3 +64,4 @@ def cleanup_deleted_files():
 
 if __name__ == '__main__':
     main()
+    syslog.syslog(syslog.LOG_INFO, "Completed media cleanup script.")
